Tortillas

+++ logica.h
@@ -1,5 +1,5 @@
- double xToFramePoint(int xp, int hres, int xmax, int xmin);
- double yToFramePoint(int yp, int vres, int ymax, int ymin);
+ double xToFramePoint(double xp, double hres, double xmax, double xmin);
+ double yToFramePoint(double yp, double vres, double ymax, double ymin);
  double getL( double xw,  double yw,  double zw,  double xe,  double ye,  double ze);
  double getXd( double xw,  double xe,  double l);
  double getYd( double yw,  double ye,  double l);
